Migrating easy to move functions out of the authenticated middleware
